cmd/charm/charmcmd: remove USSO auth
USSO auth is problematic because it only works if you
have previously authenticated with USSO auth with the
identity manager and hence already have an account record there.

Since the main reason this was present was to enable non-interactive
authentication and we now support the agent authentication protocol,
disable USSO auth in favour of that.